Move CPenaltyShootoutState traits
