+++ include/football/CPenaltyShootoutState.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "IJsonable.h"
-#include "traits/json.h"
+#include "football/traits/CPenaltyShootoutState.h"
 
 #include "types/CPenaltyShootoutState.h"
 
@@ -11,25 +11,7 @@
 
 #include "ExceptionUtils.h"
 
-namespace futsim
-{
-
-namespace football
-{
-class CPenaltyShootoutState;
-}
-
-template <> struct json_traits<football::CPenaltyShootoutState>
-{
-	//! JSON key for the class.
-	static inline constexpr std::string_view KEY = "Penalty shootout state";
-	//! JSON key for the \copybrief football::CPenaltyShootoutState::mPenalties
-	static inline constexpr std::string_view PENALTIES_KEY = "Penalties";
-	//! JSON key to indicate if the penalty corresponds to the home team.
-	static inline constexpr std::string_view HOME_TEAM_PENALTY_KEY = "Home team penalty";
-};
-
-namespace football
+namespace futsim::football
 {
 
 namespace detail
@@ -205,6 +187,4 @@ FUTSIM_CATCH_AND_RETHROW_EXCEPTION( std::invalid_argument, "Error creating the p
 
 } // detail namespace
 
-} // football namespace
-
-} // futsim namespace
+} // futsim::football namespace
